+++ changelog.md
@@ -4,6 +4,7 @@
 
 1. Fixed issue updating multiple account connections at the same time
 2. Renamed confusing `query` param key to more specific keys like (`name`, `reason`)
+3. Added `?name` query param to **GET** `/user/ban/list`
 
 ## v2.8.7